# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fire Safety & NOC Compliance Platform - API",
    "description": "Phase 1 backend (product scope Part B). Deterministic classification engine, report generator, LLM-driven dialogue manager, frontend, database persistence, and the OCR document-ingest pipeline are live; the RAG knowledge base is not yet wired in - see README.md.",
    "version": "0.1.0-phase1"
  },
  "paths": {
...
    "/users/me/case-files": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List My Case Files",
        "description": "Phase 2's starting point for Project History (\u00a7 \"Project history\"):\nthe case files this account owns, newest-updated first.\n\nPaged. Unbounded, this response grew linearly with the account's whole\nhistory - measured at 0.81 MB of JSON for 500 projects, with no ceiling\n- even though the chat rail renders eight of them.",
        "operationId": "list_my_case_files_users_me_case_files_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseFilePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/chat-titles": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List My Chat Titles",
        "description": "A title for each of this account's chats, from its first user\nmessage - what lets the sidebar tell projects apart before one is\nnamed. Scoped to case files this account owns, so it can never expose\nanother account's conversation. Sessions with no user message yet are\nsimply absent.",
        "operationId": "list_my_chat_titles_users_me_chat_titles_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatTitle"
                  },
                  "title": "Response List My Chat Titles Users Me Chat Titles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/review-queue": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List My Review Queue",
        "description": "Phase 3: every flagged case this account is responsible for - its\nown projects, plus projects shared with it for review.\n\nOLDEST FIRST, unlike every other list in this product. A chat rail is\nnewest-first because you are resuming what you were just doing; a\ncompliance queue is oldest-first because the case that has been waiting\nlongest is the one most at risk of being forgotten.\n\nApproved and rejected cases are excluded by default: they have had\ntheir answer and should stop competing for attention. Pass\ninclude_settled=true to see the whole picture.",
        "operationId": "list_my_review_queue_users_me_review_queue_get",
        "parameters": [
          {
            "name": "include_settled",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Settled"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewQueueItem"
                  },
                  "title": "Response List My Review Queue Users Me Review Queue Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/search": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Search My Projects",
        "description": "Finds a project by its name OR by something said in its conversation.\n\nSearch used to match the title only - and a title is the first message\nverbatim - so \"that project where we discussed the atrium\" still meant\nopening chats one by one.\n\nScoped to everything this account can read: its own projects, ones\nshared with it, and ones reachable through a team. The scope is\ncomputed HERE and handed to the store as a list of ids, so the search\nquery itself never has to know about access.\n\n**A substring match with no ranking** - see app/message_store.py's\n`search` for why, and what real full-text search would cost.",
        "operationId": "search_my_projects_users_me_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 200,
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectSearchHit"
                  },
                  "title": "Response Search My Projects Users Me Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/preferences": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get My Preferences",
        "description": "How this account likes to work - see app/models/preferences.py.\n\nAlways answers, defaults included: an account that has never set\nanything is not an error, and making the frontend distinguish \"no\npreferences yet\" from \"no preferences set\" would buy nothing.",
        "operationId": "get_my_preferences_users_me_preferences_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferences"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Replace My Preferences",
        "description": "Replaces this account's preferences wholesale.\n\nReturns what was actually stored, not what was sent: duplicates are\ndropped and the list is capped, so a caller that echoed its own\nrequest back would drift out of step with the server.\n\nRequires an account, obviously - there is nothing to attach a\npreference to without one. An anonymous session keeps using the\nbrowser's own storage, which is the correct place for it.",
        "operationId": "replace_my_preferences_users_me_preferences_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPreferences"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferences"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseFilePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseFile"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "CaseFilePage",
        "description": "A page of projects, with the total so the UI can say what it is not\nshowing rather than silently implying this is everything."
      },
...
      "ChatTitle": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "title"
        ],
        "title": "ChatTitle",
        "description": "A chat's title derived from what was actually said in it.\n\nDeliberately NOT a field on the Case File: a title is a presentation\ndetail of the conversation, the Case File schema is fixed, and\npersisting one would mean keeping it in step with a transcript that\nalready contains the answer. Served as a small side-lookup the chat\nrail merges over the project list it already has."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectSearchHit": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name"
          },
          "matched_in": {
            "type": "string",
            "title": "Matched In",
            "description": "'name' or 'conversation' - so the UI can say why a row is here."
          },
          "snippet": {
            "type": "string",
            "title": "Snippet",
            "description": "The matching phrase in context. Empty for a name match.",
            "default": ""
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "project_name",
          "matched_in",
          "updated_at"
        ],
        "title": "ProjectSearchHit",
        "description": "One search result."
      },
...
      "ReviewQueueItem": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name",
            "default": ""
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reason Codes",
            "description": "ReviewReasonCode values, for grouping the queue."
          },
          "reason_count": {
            "type": "integer",
            "title": "Reason Count",
            "default": 0
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "is_owner": {
            "type": "boolean",
            "title": "Is Owner",
            "description": "Whether this account owns the project or is reviewing it for someone else."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "assignment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Assignment"
              },
              {
                "type": "null"
              }
            ],
            "description": "The current assignment, if the case has one. Carried on the queue row so the list can show who owes this and by when without a request per row."
          },
          "assigned_to_me": {
            "type": "boolean",
            "title": "Assigned To Me",
            "description": "Whether the current assignment names this account - what 'my queue' filters on.",
            "default": false
          },
          "is_overdue": {
            "type": "boolean",
            "title": "Is Overdue",
            "description": "Past its due date and not yet settled. Advisory: nothing in this product acts when a due date passes, it only says so.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status",
          "is_owner",
          "updated_at"
        ],
        "title": "ReviewQueueItem",
        "description": "One row of the review queue."
      },
...
      "ReviewStatus": {
        "type": "string",
        "enum": [
          "needs_review",
          "in_review",
          "approved",
          "changes_requested",
          "rejected"
        ],
        "title": "ReviewStatus",
        "description": "Where a flagged case is in its review.\n\n`NEEDS_REVIEW` is the implicit starting state of any flagged case - it\nis never written as an event, it is simply what a case with no review\nevents yet is. Everything else is a person's recorded decision."
      },
...
      "UserPreferences": {
        "properties": {
          "pinned_session_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pinned Session Ids",
            "description": "Projects this account has pinned, most recently pinned first. Ids only - the projects themselves are read through the ordinary access checks, so a pin can never be a way to see a project you no longer have access to."
          }
        },
        "type": "object",
        "title": "UserPreferences",
        "description": "Everything this product remembers about how one person likes to work.\n\nNote what is NOT here: the active project and the session token. Those\nare per-BROWSER, not per-user - \"which project was I last looking at\"\nshould differ between the laptop and the phone, and syncing it would\nmake two open tabs fight. They stay in localStorage on purpose."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}